Add support for prefer line break at end of file. This allows us to insert code between files without affecting the source maps or deobfuscation.
This solution shifts an existing line break if it is near eof, which may result
in a line longer than the normal line length advisory limit. The only
alternative to that is to miss many eofs, which can impact the ability to place
interstitial code, or to place a break at the end of every file, which leads to
a hugely excessive number of breaks.

We could probably do better if we tracked line length across file boundaries
which we currently do not.

R=acleung,johnlenz
DELTA=149  (128 added, 0 deleted, 21 changed)


Revision created by MOE tool push_codebase.
MOE_MIGRATION=4172


git-svn-id: http://closure-compiler.googlecode.com/svn/trunk@1762 b0f006be-c8cd-11de-a2e8-8d36a3108c74